# Assignment 8/OpenAPI_Specification.yml
openapi: 3.0.3
info:
  version: 1.0.0
  title: NUEats API
  description: A food delivery application for customers, restaurant owners, and delivery persons
...
paths:
...
  /restaurant/register:
    post:
      summary: Register a new restaurant account
      tags:
        - Restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                  format: password
                address:
                  type: string
                hours:
                  type: string
              required:
                - name
                - email
                - password
                - address
                - hours
      responses:
        '201':
          description: Created
        '400':
          description: Bad request
  /restaurant/login:
    post:
      summary: Login to an existing restaurant account
      tags:
        - Restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: OK
        '400':
          description: Bad request
...